third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-for-name-in-adv-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  const name_too_long = generate_string(241, 'a');
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{name: name_too_long}]}),
    new DOMException(
      'Failed to execute \'requestDevice\' on \'Bluetooth\': ' +
      'A \'name\' or \'namePrefix\' longer than 240 bytes ' +
      'results in no devices being found, because a device can\'t ' +
      'acquire a name longer than 240 bytes.',
      'NotFoundError'),
    'Device name longer than 240');
}, 'A device name longer than 240 bytes must reject.');
</script>
